refactor: reuse flag response schema, rename api folder
