+++ apps/backend/src/api/core/docs.ts
@@ -1,5 +1,5 @@
 import EmbeddedWallets from 'api/embedded-wallets/docs'
-import FeatureFlags from 'api/feature-flags/docs'
+import FeatureFlags from 'api/general-settings/docs'
 
 import { HttpStatusCodes } from './utils/http/status-code'
