+++ apps/backend/src/api/core/routes.ts
@@ -4,7 +4,7 @@ import express, { NextFunction, Request, Response, Router } from 'express'
 import { HttpStatusCodes } from 'api/core/utils/http/status-code'
 
 import embeddedWalletsRoutes from '../embedded-wallets/routes'
-import { router as featureFlagsRoutes, adminRouter as adminFeatureFlagsRoutes } from '../feature-flags/routes'
+import { router as featureFlagsRoutes, adminRouter as adminFeatureFlagsRoutes } from '../general-settings/routes'
 
 function routes(http: express.Router): void {
   http.get('/health', (_req, res) => {
