+++ apps/backend/src/api/general-settings/docs.ts
@@ -1,7 +1,7 @@
 import { Tags } from 'api/core/utils/docs/tags'
-import CreateFeatureFlagsDocs from 'api/feature-flags/use-cases/create-feature-flag/index.docs'
-import GetFeatureFlagsDocs from 'api/feature-flags/use-cases/get-feature-flags/index.docs'
-import UpdateFeatureFlagDocs from 'api/feature-flags/use-cases/update-feature-flag/index.docs'
+import CreateFeatureFlagsDocs from 'api/general-settings/use-cases/create-feature-flag/index.docs'
+import GetFeatureFlagsDocs from 'api/general-settings/use-cases/get-feature-flags/index.docs'
+import UpdateFeatureFlagDocs from 'api/general-settings/use-cases/update-feature-flag/index.docs'
 
 export default {
   '/api/feature-flags': {
