+++ apps/backend/src/api/general-settings/use-cases/get-feature-flags/types.ts
@@ -1,18 +1,11 @@
 import { z } from 'zod'
 
 import { createResponseSchema } from 'api/core/framework/use-case/base'
+import { featureFlagSchema } from 'api/core/utils/zod'
 
 export const ResponseSchema = createResponseSchema(
   z.object({
-    flags: z
-      .object({
-        id: z.string(),
-        name: z.string(),
-        is_active: z.boolean(),
-        description: z.string().optional(),
-        metadata: z.record(z.any()).optional(),
-      })
-      .array(),
+    flags: featureFlagSchema.array(),
   })
 )
