+++ apps/backend/src/api/general-settings/use-cases/update-feature-flag/types.ts
@@ -1,10 +1,12 @@
 import { z } from 'zod'
 
 import { createResponseSchema } from 'api/core/framework/use-case/base'
+import { featureFlagSchema } from 'api/core/utils/zod'
 
 export const RequestSchema = z.object({
-  name: z.string(),
-  is_active: z.boolean(),
+  id: z.string(),
+  name: z.string().optional(),
+  is_active: z.boolean().optional(),
   description: z.string().optional(),
   metadata: z.record(z.any()).optional(),
 })
@@ -13,12 +15,7 @@ export type RequestSchemaT = z.infer<typeof RequestSchema>
 
 export const ResponseSchema = createResponseSchema(
   z.object({
-    flag: z.object({
-      name: z.string(),
-      is_active: z.boolean(),
-      description: z.string().optional(),
-      metadata: z.record(z.any()).optional(),
-    }),
+    flag: featureFlagSchema,
   })
 )
